+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Version latest (0.18.0rc?, 2019-12-05?)
 - ...
 - enh: each click group/command to have its own usage stats with its local parameters
 - enh: click `cli_core:invoke_without_command=False` 
+- enh: move isitfit command from utils to `cli.click_descendents`
+- feat: ping matomo on uncaught click exceptions. In isitfit command and group
 
 
 Version 0.17.0 (2019-12-04)